adjusted logic to reflect how input is handled
